Extract internal utilities into a separate module
