+++ mem-store.ts
@@ -1,208 +1,10 @@
 /* Copyright (c) 2010-2020 Richard Rodger and other contributors, MIT License */
 'use strict'
 
-import Assert from 'assert'
+import * as Intern from './lib/intern'
 
 let internals = {
   name: 'mem-store'
-}
-
-const intern = {
-  /* NOTE: `intern` serves as a namespace for utility functions used by
-   * the mem store.
-   */
-
-
-  // NOTE: This function is intended for use by the #save method. This
-  // function returns true when the entity argument is assumed to not yet
-  // exist in the store.
-  //
-  // In terms of code, if client code looks like so:
-  // ```
-  //   seneca.make('product')
-  //     .data$({ label, price })
-  //     .save$(done)
-  // ```
-  //
-  // - `intern.is_new` will be invoked from the #save method and return
-  // true, because the product entity is yet to be saved.
-  //
-  // The following client code will cause `intern.is_new` to return false,
-  // when invoked from the #save method, because the user entity already
-  // exists:
-  // ```
-  //   seneca.make('user')
-  //     .load$(user_id, (err, user) => {
-  //       if (err) return done(err)
-  //
-  //       return user
-  //         .data$({ email, username })
-  //         .save$(done)
-  //     })
-  // ```
-  // 
-  is_new(ent: any): boolean {
-    return null != ent && null == ent.id
-  },
-
-
-  is_upsert_requested(msg: any): boolean {
-    const { ent, q } = msg
-    return intern.is_new(ent) && Array.isArray(q.upsert$)
-  },
-
-
-  find_one_doc(entmap: any, ent: any, filter: any): any {
-    const { base, name } = ent.canon$({ object: true })
-
-    if (!(base in entmap)) {
-      return null
-    }
-
-    if (!(name in entmap[base])) {
-      return null
-    }
-
-    const coll = entmap[base][name]
-    const docs = Object.values(coll)
-    const public_entdata = ent.data$(false) 
-
-    const doc = docs.find((doc: any) => {
-      for (const fp in filter) {
-        if (fp in doc && filter[fp] === doc[fp]) {
-          continue
-        }
-
-        return false
-      }
-      
-      return true
-    })
-
-    if (!doc) {
-      return null
-    }
-
-    return doc
-  },
-
-
-  update_one_doc(entmap: any, ent: any, filter: any, new_attrs: any) {
-    const doc_to_update = intern.find_one_doc(entmap, ent, filter)
-
-    if (doc_to_update) {
-      Object.assign(doc_to_update, new_attrs)
-      return doc_to_update
-    }
-
-    return null
-  },
-
-
-  // NOTE: Seneca supports a reasonable set of features
-  // in terms of listing. This function can handle
-  // sorting, skiping, limiting and general retrieval.
-  //
-  listents(seneca: any, entmap: any, qent: any, q: any, done: any) {
-    let list = []
-
-    let canon = qent.canon$({ object: true })
-    let base = canon.base
-    let name = canon.name
-
-    let entset = entmap[base] ? entmap[base][name] : null
-    let ent
-
-    if (null != entset && null != q) {
-      if ('string' == typeof q) {
-        ent = entset[q]
-        if (ent) {
-          list.push(ent)
-        }
-      } else if (Array.isArray(q)) {
-        q.forEach(function(id) {
-          let ent = entset[id]
-          if (ent) {
-            ent = qent.make$(ent)
-            list.push(ent)
-          }
-        })
-      } else if ('object' === typeof q) {
-        let entids = Object.keys(entset)
-        next_ent: for (let id of entids) {
-          ent = entset[id]
-          for (let p in q) {
-            let qv = q[p] // query val
-            let ev = ent[p] // ent val
-
-            if (-1 === p.indexOf('$')) {
-              if (Array.isArray(qv)) {
-                if (-1 === qv.indexOf(ev)) {
-                  continue next_ent
-                }
-              } else if (null != qv && 'object' === typeof qv) {
-                // mongo style constraints
-                if (
-                  (null != qv.$ne && qv.$ne == ev) ||
-                  (null != qv.$gte && qv.$gte > ev) ||
-                  (null != qv.$gt && qv.$gt >= ev) ||
-                  (null != qv.$lt && qv.$lt <= ev) ||
-                  (null != qv.$lte && qv.$lte < ev) ||
-                  (null != qv.$in && -1 === qv.$in.indexOf(ev)) ||
-                  (null != qv.$nin && -1 !== qv.$nin.indexOf(ev)) ||
-                  false
-                ) {
-                  continue next_ent
-                }
-              } else if (qv !== ev) {
-                continue next_ent
-              }
-            }
-          }
-          ent = qent.make$(ent)
-          list.push(ent)
-        }
-      }
-    }
-
-    // Always sort first, this is the 'expected' behaviour.
-    if (q.sort$) {
-      let sf: any
-      for (sf in q.sort$) {
-        break
-      }
-
-      let sd = q.sort$[sf] < 0 ? -1 : 1
-      list = list.sort(function(a, b) {
-        return sd * (a[sf] < b[sf] ? -1 : a[sf] === b[sf] ? 0 : 1)
-      })
-    }
-
-    // Skip before limiting.
-    if (q.skip$ && q.skip$ > 0) {
-      list = list.slice(q.skip$)
-    }
-
-    // Limited the possibly sorted and skipped list.
-    if (q.limit$ && q.limit$ >= 0) {
-      list = list.slice(0, q.limit$)
-    }
-
-    // Prune fields
-    if (q.fields$) {
-      for (let i = 0; i < list.length; i++) {
-        let entfields = list[i].fields$()
-        for (let j = 0; j < entfields.length; j++) {
-          if ('id' !== entfields[j] && -1 == q.fields$.indexOf(entfields[j])) {
-            delete list[i][entfields[j]]
-          }
-        }
-      }
-    }
-
-    // Return the resulting list to the caller.
-    done.call(seneca, null, list)
-  }
 }
 
 module.exports = mem_store
@@ -213,7 +15,10 @@ module.exports.defaults = {
     'Entity of type <%=type%> with id = <%=id%> already exists.',
 }
 
-module.exports.intern = intern
+/* NOTE: `intern` serves as a namespace for utility functions used by
+ * the mem store.
+ */
+module.exports.intern = Intern
 
 function mem_store(this: any, options: any) {
   let seneca: any = this
@@ -269,7 +74,7 @@ function mem_store(this: any, options: any) {
       // if we are do a create, otherwise
       // we will do a save instead
       //
-      const is_new = intern.is_new(ent)
+      const is_new = Intern.is_new(ent)
 
       return is_new ? create_new() : do_save()
 
@@ -307,7 +112,7 @@ function mem_store(this: any, options: any) {
           mement = Object.assign(prev || {}, mement)
         }
 
-        if (intern.is_upsert_requested(msg)) {
+        if (Intern.is_upsert_requested(msg)) {
           const upsert_on = seneca.util.clean(msg.q.upsert$)
 
           if (upsert_on.length > 0) {
@@ -320,7 +125,7 @@ function mem_store(this: any, options: any) {
                 return h
               }, {})
 
-              const updated_doc = intern.update_one_doc(entmap, ent, match_by, public_entdata)
+              const updated_doc = Intern.update_one_doc(entmap, ent, match_by, public_entdata)
 
               if (updated_doc) {
                 return reply(null, ent.make$(updated_doc))
@@ -333,7 +138,7 @@ function mem_store(this: any, options: any) {
 
         seneca.log.debug(function() {
           return [
-            'save/' + (intern.is_new(msg.ent) ? 'insert' : 'update'),
+            'save/' + (Intern.is_new(msg.ent) ? 'insert' : 'update'),
             ent.canon$({ string: 1 }),
             mement,
             desc,
@@ -390,7 +195,7 @@ function mem_store(this: any, options: any) {
       let qent = msg.qent
       let q = msg.q
 
-      return intern.listents(this, entmap, qent, q, function(this: any, err: any, list: any[]) {
+      return Intern.listents(this, entmap, qent, q, function(this: any, err: any, list: any[]) {
         let ent = list[0] || null
 
         this.log.debug(function() {
@@ -405,7 +210,7 @@ function mem_store(this: any, options: any) {
       let qent = msg.qent
       let q = msg.q
 
-      return intern.listents(this, entmap, qent, q, function(this: any, err: any, list: any[]) {
+      return Intern.listents(this, entmap, qent, q, function(this: any, err: any, list: any[]) {
         this.log.debug(function() {
           return [
             'list',
@@ -430,7 +235,7 @@ function mem_store(this: any, options: any) {
       // default false
       let load = q.load$ === true
 
-      return intern.listents(seneca, entmap, qent, q, function(err: Error, list: any[]) {
+      return Intern.listents(seneca, entmap, qent, q, function(err: Error, list: any[]) {
         if (err) {
           return reply(err)
         }
